Fixed errors in Chrome Console + New Color Scheme
There were like 59 errors, 54 were related to the sidebars because of how I referenced them and 1 was due to some code not really being used in App.vue. The other 4 errors are related to me testing SignUp.vue component so it isn't a concern.

And also, I've changed up the color scheme for the time being. Feedback from Glory will be welcome!