Sign Up works with POST to Songstress + Vuelidate Package
So Sign Up should be working, only know from the console logs I get when Songstress is running. Also, I added the vuelidate package for forms but it is not implemented yet...gonna get to it shouldnt take much time. More impt to get Sign Up/Sign In as dialog actions first imo